Rename main package from optimizer to cloud_resource_matcher
Optimizer is too generic.
