+++ benches/utils/data_generation.py
@@ -1,7 +1,7 @@
 """Utility functions to generate the benchmark data."""
-from optimizer.modules.base.data import CloudService, BaseData, CloudResource
-from optimizer.modules.network import NetworkData
-from optimizer.modules.network.data import Location
+from cloud_resource_matcher.modules.base.data import CloudService, BaseData, CloudResource
+from cloud_resource_matcher.modules.network import NetworkData
+from cloud_resource_matcher.modules.network.data import Location
 
 
 def generate_base_data(cr_count: int, cs_count: int, cs_count_per_cr: int) -> BaseData:
