+++ cloud_resource_matcher/modules/base/extract_solution.py
