+++ cloud_resource_matcher/modules/multi_cloud/validate.py
@@ -1,7 +1,7 @@
 """Implementation of the validation step for the multi cloud module."""
 from optiframe.framework.tasks import ValidateTask
 
-from optimizer.modules.base import BaseData
+from cloud_resource_matcher.modules.base import BaseData
 from .data import MultiCloudData
