+++ test/modules/network/test_validate.py
@@ -1,9 +1,9 @@
 """Tests for the validation step of the network module."""
 import pytest
 
-from optimizer.modules.base import BaseData
-from optimizer.modules.network import NetworkData
-from optimizer.modules.network.validate import ValidateNetworkTask
+from cloud_resource_matcher.modules.base import BaseData
+from cloud_resource_matcher.modules.network import NetworkData
+from cloud_resource_matcher.modules.network.validate import ValidateNetworkTask
 
 
 def test_should_not_raise_error_for_valid_data() -> None:
