+++ test/modules/performance/test_build_mip.py
@@ -2,9 +2,9 @@
 from optiframe import Optimizer
 from pulp import LpMinimize
 
-from optimizer.modules.base import BaseData, base_module
-from optimizer.modules.performance import PerformanceData, performance_module
-from optimizer.modules.service_limits import service_limits_module, ServiceLimitsData
+from cloud_resource_matcher.modules.base import BaseData, base_module
+from cloud_resource_matcher.modules.performance import PerformanceData, performance_module
+from cloud_resource_matcher.modules.service_limits import service_limits_module, ServiceLimitsData
 from test.framework import Expect
